agent-server/requirements.txt
# BrowserAgent — Python Dependencies
# Install: pip install -r requirements.txt

fastapi>=0.115.0,<1.0.0
uvicorn[standard]>=0.32.0,<1.0.0
websockets>=14.0,<15.0
pydantic>=2.10.0,<3.0.0
pydantic-settings>=2.7.0,<3.0.0
httpx>=0.28.0,<1.0.0
google-generativeai>=0.8.0,<1.0.0
numpy>=2.2.0,<3.0.0
hnswlib>=0.8.0,<1.0.0
aiosqlite>=0.21.0,<1.0.0
Pillow>=11.0.0,<12.0.0
python-dotenv>=1.0.0,<2.0.0
